refactor and improve collector logic
stderr and stdout (and worker meta data) are automatically collected by a
background process responsible for maintaining the output order.  This
approach dumps collected output as soon as it is available, rather than
waiting until all chunk processing has finished.

There are currently no predefined collectors, since stdout and stderr are
always collected.  This was to prove the concept, rather than being the
final intention.  Ultimately, all collectors will be run by the background
"collector manager" process, and stdout and stderr will be treated just like
any user-defined collector.

This new collector behaviour required all chunk input files to be
sequentially numbered, which implied that all work should now be done in a
clean and temporary directory.  This meant a few changes in order to account
for the slightly different locations of created and output files.  It also
means that relative paths shouldn't be used for the command or any of its
arguments, because ll-pipe changes into the temp directory while it is
working.  Using a temporary directory meant creating no temporary filenames
with mktemp because we can presume to have complete control over what goes
into that directory.  It also means clean up for the --clean-all is
extremely simple.