Implement predefined collectors.
+stdout, +stderr, and +meta are predefined to collect successful chunks'
.out, .err, and .meta files.

If no collectors are given, +stdout and +stderr are enabled.  If ANY
collectors are given, the user has to collect .out and .err themselves if
they want it (although they can specify the +stdout and/or +stderr to do
it).

The .meta collector is always enabled (although without --meta-outfile-json,
the collected meta data is thrown away).