Collect metadata from collectors.
When each collector is kicked off, it is invoked via /usr/bin/time so we can
collect meta data from its run.  This metadata is appended to the
--meta-outfile-json file, if given.

Of note, the change in timeCommand() to remove --output from the list of
options is due to the fact that the way the code was before, if the command
passed to timeCommand had a long-option that started with --output, it would
also be consumed (due to passing -E to zparseopts), unless -- was used to
delimit options from the command arguments.  That change was made also to
remove all possible ambiguity.