Better worker signal handling.
Don't set numAttempts to maximum in order to force an exit from the retry
loop because that makes it look like suddenly the command failed, when it
technically didn't.  Use a flag to indicate if a signal was received.

Also output a status message if the worker is exiting because of a signal.