Report jsonify failures to caller.
Discovered that if your env is missing the JSON::XS perl module, metadata
post-processing will not be done, stderr for each chunk will contain the
failure messages from perl saying it can't find the module, but the chunks
will still succeed.

This change makes jsonifying failures be fatal for the chunk.

I'm going to remove the JSON stuff anyway because the metadata can be
post-processed from whatever form it gets written in as long as I choose
something sane to parse.